# doc/api.yaml
openapi: 3.0.0
info:
  title: WasaPhoto - OpenAPI 3.0
  description: | 
    This is the api specification for WasaPhoto, an educational purpose based project, 
    to keep in touch with your friends by sharing photos of special moments.  
    - [The github private project repository](https://github.com/luigidannibale/Wasa)
    
    Functional design specifications: 
    
    Each user will be presented with a stream of photos (images) in reverse chronological order, with
    information about when each photo was uploaded (date and time) and how many likes and comments
    it has. The stream is composed by photos from “following” (other users that the user follows). Users
    can place (and later remove) a “like” to photos from other users. Also, users can add comments to any
    image (even those uploaded by themself). Only authors can remove their comments.

    Users can ban other users. If user Alice bans user Eve, Eve won’t be able to see any information about
    Alice. Alice can decide to remove the ban at any moment.
    
    Users will have their profiles. The personal profile page for the user shows: the user’s photos (in reverse
    chronological order), how many photos have been uploaded, and the user’s followers and following.
    Users can change their usernames, upload photos, remove photos, and follow/unfollow other users.
    Removal of an image will also remove likes and comments.
    
    A user can search other user profiles via username.
    
...

...
  contact:
...
security:
  - UserAuth: []
paths:        
...
    /users/{userID}/followed/{followedID}:
      parameters:
      - $ref : "#/components/parameters/userID"
      - $ref : "#/components/parameters/followedID"
      
      delete:
        operationId: unfollowUser
        tags:
          - User Interactions
          
        summary: Unfollows a user
        description: |
          The actual user (userID) unfollows the given user (followedID).          
        responses:
          "200": 
            $ref : "#/components/responses/OK"
          "400":
            $ref : "#/components/responses/BadRequest"
          "401":
            $ref : "#/components/responses/Unauthorized"
          "403":
            $ref : "#/components/responses/Forbidden"
          "404":
            $ref : "#/components/responses/NotFound"
          "500": 
            $ref : "#/components/responses/InternalServerError"
    
    #getBannedList    - LastMod : 11/12 
    #banUser          - LastMod : 11/12 
...
    /users/{userID}/banned/{bannedID}:
      parameters:
      - $ref : "#/components/parameters/userID"
      - $ref : "#/components/parameters/bannedID"

      delete:
        operationId: unbanUser
        tags:
          - User Interactions
          
        summary: Unbans a user
        description: |
          The actual user (user ID) unbans the given user (banned ID).           
        responses:
          "200": 
            $ref : "#/components/responses/OK"
          "400":
            $ref : "#/components/responses/BadRequest"
          "401":
            $ref : "#/components/responses/Unauthorized"
          "403":
            $ref : "#/components/responses/Forbidden"
          "404":
            $ref : "#/components/responses/NotFound"
          "500": 
            $ref : "#/components/responses/InternalServerError"
    
    #uploadPhoto       - LastMod : 11/12 - Status : Working BanSafe
...
    /photos/{photoID}:
      parameters:
        - $ref : "#/components/parameters/photoID"
      get:
        operationId: getPhoto
        tags:
          - Photos Interactions
          
        description: 
          Gets a photo by the id.
        summary: Gets a photo
        responses:
          "200":
            description: Successful, returning the photo
            content: 
              application/json:
                schema:
                  $ref : "#/components/schemas/Photo"            
          "400":
            $ref : "#/components/responses/BadRequest" 
          "401":
            $ref : "#/components/responses/Unauthorized"
          "404":
           $ref : "#/components/responses/NotFound" 
          "500":
            $ref : "#/components/responses/InternalServerError" 
      delete:
        operationId: deletePhoto
        tags:
          - Photos 
          
        description: |
          Deletes a photo by the id.          
        summary: Deletes a photo
        responses:
          "200":
            $ref : "#/components/responses/OK" 
          "400":
            $ref : "#/components/responses/BadRequest" 
          "401":
            $ref : "#/components/responses/Unauthorized" 
          "403":
           $ref : "#/components/responses/Forbidden"
          "404":
           $ref : "#/components/responses/NotFound" 
          "500":
            $ref : "#/components/responses/InternalServerError"     
    
    #likePhoto   - LastMod : 11/12 - Status : Working BanSafe
    #getLike     - LastMod : 11/12 - Status : Working BanSafe
    #unlikePhoto - LastMod : 11/12 - Status : Working BanSafe
    /photos/{photoID}/likes:
      parameters:
        - $ref : "#/components/parameters/photoID"
      put:
        operationId: likePhoto
        tags: 
          - Photos Interactions
          
        summary: Puts a like to a photo
        description: |
          Takes a userID and creates a like of that user to the photo, 
          if such user has already liked the photo, nothing happens.        
        responses:
          "200":
            $ref : "#/components/responses/OK" 
          "201":
            $ref : "#/components/responses/Created" 
          "400":
            $ref : "#/components/responses/BadRequest"
          "401":
            $ref : "#/components/responses/Unauthorized"
          "404":
            $ref : "#/components/responses/NotFound"            
          "500":
            $ref : "#/components/responses/InternalServerError"
      get:
        operationId: getLike
        tags:
          - Photos Interactions
          
        description: 
          Gets a like
        summary: Gets a like
        responses:
          "200":
            description: Successful, returning the like
            content: 
              application/json:
                schema:
                  $ref : "#/components/schemas/Message"
          "400":
            $ref : "#/components/responses/BadRequest" 
          "401":
            $ref : "#/components/responses/Unauthorized"
          "404":
           $ref : "#/components/responses/NotFound" 
          "500":
            $ref : "#/components/responses/InternalServerError" 
      delete:
        operationId: unlikePhoto
        tags:
          - Photos Interactions
          
        summary: Deletes a like to the photo
        description: |
          Takes a userID, if it matches the user that has put the like, the like is deleted,
          otherwise nothing happens.
        responses:
          "200":
            $ref : "#/components/responses/OK" 
          "400":
            $ref : "#/components/responses/BadRequest" 
          "401":
            $ref : "#/components/responses/Unauthorized" 
          "403":
            $ref : "#/components/responses/Forbidden" 
          "404":
            $ref : "#/components/responses/NotFound" 
          "500":
            $ref : "#/components/responses/InternalServerError" 
    
    #getLikersList - LastMod : 11/12 - Status : Working BanSafe
...
    /photos/{photoID}/comments/{commentID}:
      parameters: 
        - $ref : "#/components/parameters/photoID"
        - $ref : "#/components/parameters/commentID"
      
      get:
        operationId: getComment
        tags: 
          - Photos Interactions
          
        summary: Gets the comment details
        description: Takes a comment id and returns the comment
        responses:
          "200":
            description:  Success, returning the comment
            content:
              application/json:
                schema:
                   $ref : "#/components/schemas/Message"
          "400":
            $ref : "#/components/responses/BadRequest" 
          "401":
            $ref : "#/components/responses/Unauthorized"
          "404":
            $ref : "#/components/responses/NotFound" 
          "500":
            $ref : "#/components/responses/InternalServerError"        
      delete:
        operationId: uncommentPhoto
        tags:
          - Photos Interactions
          
        summary: Deletes a comment to the photo
        description: |
          Deletes a comment to the photo, using both photoID and commentID.
        responses:
          "200":
            description: Success, comment deleted
            content:
              application/json:
                schema:
                   $ref : "#/components/schemas/Message"
          "400":
            $ref : "#/components/responses/BadRequest"
          "401":
            $ref : "#/components/responses/Unauthorized"
          "403":
            $ref : "#/components/responses/Forbidden"
          "404":
            $ref : "#/components/responses/NotFound" 
          "500":
            $ref : "#/components/responses/InternalServerError"
components:  
  securitySchemes: # LastMod - 2/12
    UserAuth: 
      type: http
      scheme: bearer
  parameters: # LastMod - 28/11
    userID: 
      name: userID
      in: path
      description: Id of the user
      required: true
      schema:
        $ref : "#/components/schemas/User/properties/id"
    followedID: 
      name: followedID
      in: path
      description: Id of the followed user
      required: true
      schema:
        $ref : "#/components/schemas/User/properties/id"
    bannedID: 
      name: bannedID
      in: path
      description: Id of the banned user
      required: true
      schema:
        $ref : "#/components/schemas/User/properties/id"
    photoID: 
      name: photoID
      in: path
      description: ID of the photo
      required: true
      schema:
        $ref : "#/components/schemas/Photo/properties/id"    
    commentID: 
      name: commentID
      in: path
      description: ID of the comment
      required: true
      schema:
        $ref : "#/components/schemas/Comment/properties/id"    
  schemas:  #LastMod - 11/12
...
    Timestamp: 
      type: string
      description: a timestamp composed of date, hour, minutes, seconds
      pattern: "[0-9]{2} [0-9]{2} [0-9]{2}:[0-9]{2}:[0-9]{2}[A|P]M '[0-9]{2} .[0-9]{4}"
      example: "12 11 03:08:38PM '23 +0100"
    Photo: 
      type: object 
      description: a photo composed of id, image, caption, upload timestamp
      properties:
        id:
          description: id of the photo
          type: integer
          example: 12
          readOnly: true
        image:          
          description: image of the photo
          type: string
          format: binary          
          minLength: 0
          maxLength: 40960          
        caption:
          description: caption of the photo
          type: string  
          example: Look what a beautiful photo
          pattern: ".{1,100}$"
          minLength: 1
          maxLength: 100
        uploadTimestamp:
          $ref: "#/components/schemas/Timestamp"
...
    Message:       
      description: a simple text message
      type: string
      example: "This is the outcome"
      pattern: ".{0,1024}"
      minLength: 0
      maxLength: 1024
...
  responses: # LastMod - 2/12
    OK: 
      description: The request has been completed successfully
      content: 
        application/json:
          schema:
            $ref : "#/components/schemas/Message"
    Created: 
      description: The request has been completed successfully, resulting in a creation
      content: 
        application/json:
          schema:
            $ref : "#/components/schemas/Message"
    BadRequest: 
      description: The request is bad asked
      content: 
        application/json:
          schema:
            $ref : "#/components/schemas/Message"
    Unauthorized: 
      description: This request is unauthorized
      content: 
        application/json:
          schema:
            $ref : "#/components/schemas/Message"    
    Forbidden: 
      description: Resource(s) not found
      content: 
        application/json:
          schema:
            $ref : "#/components/schemas/Message"
    NotFound: 
      description: Resource(s) not found
      content: 
        application/json:
          schema:
            $ref : "#/components/schemas/Message"    
...
    InternalServerError: 
      description: Internal server error has occurred
      content: 
        application/json:
          schema:
            $ref : "#/components/schemas/Message"